feat(post): add functionality to update post url
Adds a new route to update the URL of a post.
This allows users to change the URL of their posts after they have
been uploaded, as long as they are in pending state.
This functionality is also protected by the userAuth middleware, ensuring
only authenticated users can update their posts.
The user's postUploadList and the corresponding PostUploadModel in the
database are also updated to reflect the changes.